% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(
  data,
  indicateur = NULL,
  correction = NULL,
  freq = NULL,
  activite = NULL
)
}
\arguments{
\item{data}{A data-frame}

\item{indicateur}{Subject of study}

\item{correction}{CVS or BRUT}

\item{freq}{M, T or A (mensuelle, etc.)}

\item{activite}{filtrer par rapport à colonne dont le libellé est fourni}
}
\value{
A filtered data-frame
}
\description{
filtering data
}
