% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download series data from Insee BDM API}
\usage{
get_data(
  domaine,
  indicateur = NULL,
  correction = NULL,
  freq = NULL,
  activite = NULL
)
}
\arguments{
\item{domaine}{Domain of the serie}

\item{indicateur}{Subject of study}

\item{correction}{CVS or BRUT}

\item{freq}{M, T or A (mensuelle, etc.)}

\item{activite}{filtrer par rapport à colonne dont le libellé est fourni}
}
\value{
A data-frame
}
\description{
downloading data
}
