% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_with_id.R
\name{get_data_with_id}
\alias{get_data_with_id}
\title{Download series data with IDBANK}
\usage{
get_data_with_id(id)
}
\arguments{
\item{id}{IDBANK}
}
\value{
A data-frame
}
\description{
downloading data with an idbank
}
