% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{plotter}
\usage{
plotter(
  data,
  groupby,
  title = "",
  xlabel = "",
  ylabel = "",
  top_margin = 0,
  title_y_position = 1.5,
  length_labels = 30
)
}
\arguments{
\item{data}{Data frame}

\item{groupby}{Group to multiplot}

\item{title}{Title of plot}

\item{xlabel}{xlabel}

\item{ylabel}{ylabel}

\item{top_margin}{top margin above plot}

\item{title_y_position}{top title position}

\item{length_labels}{length of labels while hovering}
}
\value{
A plot
}
\description{
generating data
}
